修正 總表查詢 SQL 語法, 修正 javascript int 與 float 小數點位數造成的計算錯誤

+++ src/kaoins6/FreightForm/FFShareFun.js
@@ -235,8 +235,8 @@ function UpdateTotalPieceAndVolume() {
             bSkipAnyPiece = true
         }
 
-        nPieceSum += parseInt(nPiece);
-        fVolumeSum += parseInt(document.form[18 + i * nFieldDiff].value);
+        nPieceSum += parseFloat(nPiece);
+        fVolumeSum += parseFloat(document.form[18 + i * nFieldDiff].value);
     }
     document.form.StoreSum_Piece.value = nPieceSum;
